Development

+++ src/components/AdminDashboard/Content&Resources/content/bank/MCQBank/McqBankCardForAdmin.tsx
@@ -1,8 +1,9 @@
 import CommonButton from "@/common/button/CommonButton";
+import AlertDialogBox from "@/common/custom/AlertDialogBox";
 import { toBerhanTime } from "@/help/help";
 import { useDeleteFlashCardBankMutation } from "@/store/features/adminDashboard/ContentResources/flashCard/flashCardSlice";
 import { useDeleteMcqBankApiMutation } from "@/store/features/adminDashboard/ContentResources/MCQ/mcqApi";
-import { SingleMcqBank } from "@/store/features/adminDashboard/ContentResources/MCQ/type/allContent";
+import { SingleMcqBank } from "@/store/features/adminDashboard/ContentResources/MCQ/types/allContent";
 import { useAppSelector } from "@/store/hook";
 import { RootState } from "@/store/store";
 import { FC } from "react";
@@ -19,14 +20,7 @@ const McqBankCardForAdmin: FC<Props> = ({ data, setMcqBankId }) => {
   const [deleteFlashCardBank, { isLoading: isDeleting }] =
     useDeleteFlashCardBankMutation();
 
-  // useDeleteFlashCardBankMutation,
-  // useDeleteSingleFlashCardMutation,
-
   const handleDelete = async (id: string) => {
-    const confirmDelete = window.confirm(
-      "Are you sure you want to delete this MCQ bank?"
-    );
-    if (!confirmDelete) return;
     if (contentType === "MCQ") {
       await deleteMcqBankApi(id);
     }
@@ -85,13 +79,10 @@ const McqBankCardForAdmin: FC<Props> = ({ data, setMcqBankId }) => {
       </div>
 
       <div className="p-4 bg-gray-50 flex justify-between border-t border-border ">
-        <CommonButton
-          disabled={isLoading}
-          onClick={() => handleDelete(data._id)}
-          className=" bg-red-500 !text-white"
-        >
-          {isLoading || isDeleting ? "Deleting..." : "Delete"}
-        </CommonButton>
+        <AlertDialogBox
+          action={() => handleDelete(data._id)}
+          isLoading={isDeleting || isLoading}
+        />
         <CommonButton
           onClick={() => setMcqBankId(data._id)}
           className=" bg-blue-500 !text-white"
